various cleanup, dedupe
